4.0 dev

See merge request dt-insight-engine/DAGScheduleX!269

+++ engine-master/engine-master-server/src/test/java/com/dtstack/engine/master/impl/ClusterServiceTest.java
@@ -195,7 +195,7 @@ public class ClusterServiceTest extends AbstractTest {
 
         //删除组件
         try {
-            componentService.delete(Lists.newArrayList(hdfsComponent.getId()));
+            componentService.delete(Lists.newArrayList(hdfsComponent.getId().intValue()));
         } catch (Exception e) {
             if (e instanceof RdosDefineException) {
                 RdosDefineException rdosDefineException = (RdosDefineException) e;
